move sensor.cli commands to cli
