--- src/pms/sensor/cli.py
@@ -1,76 +1,0 @@
-from datetime import datetime
-from enum import Enum
-from pathlib import Path
-from typing import Optional
-
-from typer import Argument, Context, Option, echo
-
-from pms import logger
-from pms.sensor import MessageReader
-
-
-class Format(str, Enum):
-    csv = "csv"
-    pm = "pm"
-    num = "num"
-    raw = "raw"
-    cf = "cf"
-    atm = "atm"
-    hcho = "hcho"
-    bme = "bme"
-    bsec = "bsec"
-    hex = "hexdump"
-
-
-def serial(
-    ctx: Context,
-    format: Optional[Format] = Option(None, "--format", "-f", help="formatted output"),
-    decode: Optional[Path] = Option(None, help="decode captured messages"),
-):
-    """Read sensor and print measurements"""
-    reader = ctx.obj["reader"]
-    if decode:
-        reader = MessageReader(decode, reader.sensor, reader.samples)
-    with reader:
-        if format == "hexdump":
-            for n, raw in enumerate(reader(raw=True)):
-                echo(raw.hexdump(n))
-        elif format:
-            if format == "csv":
-                obs = next(reader())
-                echo(f"{obs:header}")
-            for obs in reader():
-                echo(f"{obs:{format}}")
-        else:  # pragma: no cover
-            for obs in reader():
-                echo(str(obs))
-
-
-def csv(
-    ctx: Context,
-    capture: bool = Option(False, "--capture", help="write raw messages instead of observations"),
-    overwrite: bool = Option(False, "--overwrite", help="overwrite file, if already exists"),
-    path: Path = Argument(Path(), help="csv formatted file", show_default=False),
-):
-    """Read sensor and print measurements"""
-    if path.is_dir():  # pragma: no cover
-        path /= f"{datetime.now():%F}_pypms.csv"
-    mode = "w" if overwrite else "a"
-    logger.debug(f"open {path} on '{mode}' mode")
-    with ctx.obj["reader"] as reader, path.open(mode) as csv:
-        sensor_name = reader.sensor.name
-        if not capture:
-            logger.debug(f"capture {sensor_name} observations to {path}")
-            # add header to new files
-            if path.stat().st_size == 0:
-                obs = next(reader())
-                csv.write(f"{obs:header}\n")
-            for obs in reader():
-                csv.write(f"{obs:csv}\n")
-        else:
-            logger.debug(f"capture {sensor_name} messages to {path}")
-            # add header to new files
-            if path.stat().st_size == 0:
-                csv.write("time,sensor,hex\n")
-            for raw in reader(raw=True):
-                csv.write(f"{raw.time},{sensor_name},{raw.hex}\n")
